% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn_tabu_gen_1.R
\name{Batch_Net}
\alias{Batch_Net}
\title{Generates a subgraph from query nodes}
\usage{
Batch_Net(bn_output, Node_ids, relationship_type = "nbr")
}
\arguments{
\item{bn_output}{R object output from bn_tabu_gen function}

\item{Node_ids}{vector containing the nodes for subgraph generation.}

\item{relationship_type}{the relationship to be returned for the 
specified query nodes.
The options are "mb", "nbr", "parents", "children". Default 
setting is "nbr".}
}
\value{
a subgraph containg the selected nodes and relationships.
}
\description{
Generates a subgraph from query nodes
}
