% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmicR_viz.R
\name{Gmic_viz}
\alias{Gmic_viz}
\title{Visualized network}
\usage{
Gmic_viz(Auto_WGCNA_Output, Filter_unconnected_ME = TRUE)
}
\arguments{
\item{Auto_WGCNA_Output}{R object with GMIC bayesian network}

\item{Filter_unconnected_ME}{a logical value. If TRUE, the default,
unconnected modules will be removed from the final network. If FALSE,
all modules will be shown.}
}
\value{
a shiny object for network visualization.
}
\description{
Visualized network
}
\examples{
GMIC_Final_dir<-system.file("extdata", "GMIC_Final.Rdata", 
package = "GmicR", mustWork = TRUE)
load(GMIC_Final_dir)
if(interactive()){
Gmic_viz(GMIC_Final)}
}
