% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_annotation.R
\name{GO_Module_NameR}
\alias{GO_Module_NameR}
\title{GO enrichment for module names}
\usage{
GO_Module_NameR(Auto_WGCNA_OUTPUT, cutoff_size = 100)
}
\arguments{
\item{Auto_WGCNA_OUTPUT}{object returned by GSEAGO_Builder function}

\item{cutoff_size}{integer indication the GO size cut off. Default is 100}
}
\value{
a data.frame listing the GO name for each module. Additionally, 
a table similar to the output of Query_Prep function is returned, appended 
with module GO names.
}
\description{
GO enrichment for module names
}
\examples{
GMIC_Builder_dir<-system.file("extdata", "GMIC_Builder.Rdata", 
package = "GmicR", mustWork = TRUE)
load(GMIC_Builder_dir)
GMIC_Builder<-GO_Module_NameR(GMIC_Builder, cutoff_size = 100)
}
