% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_WGCNA.R
\name{xCell_loader}
\alias{xCell_loader}
\title{Scales and centers data by sample/row in preparation for discretization}
\usage{
xCell_loader(File = NULL)
}
\arguments{
\item{File}{the name of the text file generated by xCell that 
contains the
cell signature scores.}
}
\value{
xCell signatures scaled and centered by sample. For GMIC,
ImmuneScore, StromaScore, and MicroenvironmentScore are removed.
}
\description{
Scales and centers data by sample/row in preparation for discretization
}
\examples{
file_dir<-system.file("extdata", "IRIS_xCell_sig.txt", 
package = "GmicR", mustWork = TRUE)
Xcell_sig<-xCell_loader(file_dir)
plot(Xcell_sig$Bcells)
}
