% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_annotation.R
\name{GSEAGO_Builder}
\alias{GSEAGO_Builder}
\title{GO enrichment for module names}
\usage{
GSEAGO_Builder(Auto_WGCNA_OUTPUT, species = "Homo sapiens",
  ontology = "BP", colname_correct = TRUE)
}
\arguments{
\item{Auto_WGCNA_OUTPUT}{output from Auto_WGCNA function.}

\item{species}{either 'Homo sapiens' (default) or 'Mus musculus'.}

\item{ontology}{string either 'BP'(Biological Process; default), 
'CC'(Cellular Component),
or 'MF' (Molecular Function).}

\item{colname_correct}{a boolean value. If TRUE (default), "." in gene 
names will be replaced
with "-". This corrects a name change that is induced by R when creating a 
data.frame. If FALSE,
no changes will be made.}
}
\value{
Lists with gene ontology 
enrichment analysis, performed using GOstats,
for each module.
}
\description{
GO enrichment for module names
}
\note{
gene names must be official gene symbol
GOname work flow inspired by: 
\href{https://www.bioconductor.org/packages/devel/bioc/vignettes/
CVE/inst/doc/WGCNA_from_TCGA_RNAseq.html}{Module_GO_Names}.
}
